@@ -1,5 +1,5 @@
 {
-  "name": "udemy-nodejs-angular2",
+  "name": "unsafe-space",
   "version": "1.0.0",
   "private": true,
   "scripts": {
@@ -8,55 +8,56 @@
     "build:prod": "del-cli public/js/app && ngc -p tsconfig.aot.json && ngc -p tsconfig.aot.json && webpack --config webpack.config.prod.js --progress --profile --bail && del-cli 'public/js/app/**/*.js' 'public/js/app/**/*.js.map' '!public/js/app/bundle.js' '!public/js/app/*.chunk.js' 'assets/app/**/*.ngfactory.ts' 'assets/app/**/*.shim.ts' 'assets/app/**/*.ngsummary.json' 'assets/app/**/*.ngstyle.ts'"
   },
   "dependencies": {
-    "@angular/animations": "^4.3.5",
-    "@angular/common": "^4.3.5",
-    "@angular/compiler": "^4.3.5",
-    "@angular/compiler-cli": "^4.3.5",
-    "@angular/core": "^4.3.5",
-    "@angular/forms": "^4.3.5",
-    "@angular/http": "^4.3.5",
-    "@angular/platform-browser": "^4.3.5",
-    "@angular/platform-browser-dynamic": "^4.3.5",
-    "@angular/platform-server": "^4.3.5",
-    "@angular/router": "^4.3.5",
-    "@angular/upgrade": "^4.3.5",
+    "@angular/animations": "^4.4.7",
+    "@angular/common": "^4.4.7",
+    "@angular/compiler": "^4.4.7",
+    "@angular/compiler-cli": "^4.4.7",
+    "@angular/core": "^4.4.7",
+    "@angular/forms": "^4.4.7",
+    "@angular/http": "^4.4.7",
+    "@angular/platform-browser": "^4.4.7",
+    "@angular/platform-browser-dynamic": "^4.4.7",
+    "@angular/platform-server": "^4.4.7",
+    "@angular/router": "^4.4.7",
+    "@angular/upgrade": "^4.4.7",
     "@types/socket.io-client": "^1.4.32",
     "bcrypt-nodejs": "0.0.3",
     "bcryptjs": "^2.4.3",
-    "bluebird": "^3.5.0",
+    "bluebird": "^3.5.1",
     "body-parser": "~1.15.2",
     "connect-flash": "^0.1.1",
     "cookie-parser": "~1.4.3",
-    "core-js": "^2.5.0",
+    "core-js": "^2.5.6",
     "debug": "~2.2.0",
-    "ejs": "^2.5.7",
+    "ejs": "^2.6.1",
     "express": "~4.14.0",
-    "express-jwt": "^5.3.0",
+    "express-jwt": "^5.3.1",
     "hbs": "~3.1.0",
     "jsonwebtoken": "^7.4.3",
-    "mongoose": "^4.11.7",
-    "mongoose-unique-validator": "^1.0.5",
+    "mongoose": "^4.13.13",
+    "mongoose-unique-validator": "^1.0.6",
     "morgan": "~1.6.1",
-    "reflect-metadata": "^0.1.3",
-    "rxjs": "^5.4.3",
+    "reflect-metadata": "^0.1.12",
+    "rxjs": "^5.5.10",
     "serve-favicon": "~2.3.0",
-    "socket.io": "^2.0.4",
-    "socket.io-client": "^2.1.0",
+    "socket.io": "^2.1.1",
+    "socket.io-client": "^2.1.1",
     "socketio-jwt": "^4.5.0",
-    "zone.js": "^0.8.16"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
+    "@angular/cli": "^6.0.3",
     "@types/core-js": "0.9.36",
-    "@types/node": "^6.0.87",
+    "@types/node": "^6.0.111",
     "angular-router-loader": "^0.5.0",
     "angular2-template-loader": "^0.5.0",
-    "awesome-typescript-loader": "^3.2.3",
+    "awesome-typescript-loader": "^3.5.0",
     "del-cli": "^0.2.0",
     "html-loader": "^0.4.4",
     "raw-loader": "^0.5.1",
-    "ts-loader": "^2.3.3",
-    "typescript": "^2.4.2",
+    "ts-loader": "^2.3.7",
+    "typescript": "^2.8.3",
     "webpack": "^2.2.1",
-    "webpack-merge": "^4.1.0"
+    "webpack-merge": "^4.1.2"
   }
 }
