@@ -4,6 +4,7 @@ import { ShoutOutsComponent } from "./ShoutOuts/shoutouts.component";
 import { AuthenticationComponent } from "./Auth/authentication.component";
 import { ProfileComponent } from "./profile/profile.component";
 import { ChatComponent } from "./chat/chat.component";
+import { LobbyComponent } from "./lobby/lobby.component";
 
 import { HomeComponent } from "./home.component";
 
@@ -12,8 +13,9 @@ const APP_ROUTES: Routes = [
     { path: 'shoutouts', component: ShoutOutsComponent },
     { path: 'auth', component: AuthenticationComponent, loadChildren: './Auth/auth.module#AuthModule' },
     { path: 'home', component: HomeComponent},
-    { path: 'profile', component: ProfileComponent },
-    { path: 'chat', component: ChatComponent }
+    { path: 'profile/:id', component: ProfileComponent },
+    { path: 'chat', component: ChatComponent },
+    { path: 'lobby/:id', component: LobbyComponent }
 ];
 
 export const routing = RouterModule.forRoot(APP_ROUTES);
