@@ -1,6 +1,7 @@
 export class ChatMessage {
     constructor(public message: string,
-                public username: string
+                public username: string,
+                public chatId:string
         ){
     }
 }
