@@ -56,6 +56,10 @@ module.exports = function(server) {
     chatNsp.on('connection',
         function(socket) {
             console.log('a user connected');
+            User.findById(socket.id).exec()
+            .then( (user) => {
+
+            })
 
             socket.on('changeRoom', function(room) {
                 changeRoom(socket, room);
@@ -63,10 +67,11 @@ module.exports = function(server) {
 
             socket.on('chatMessage', function(message) {
                 message.username = socket.decoded.user.username;
+                console.log('message: ' + message);
+                console.log('Chatroom: ' + socket.decoded.user.data);
                 chatNsp.in('default').emit('chatMessage', message);
                 message.rooms = socket.rooms;
                 delete message.rooms[socket.id];
-                console.log(message);
             });
 
             socket.on('getGames', function() {
@@ -76,25 +81,25 @@ module.exports = function(server) {
                     for(var i =0; i<games.length; i++) {
                         clientGames.push({
                             id: games[i]._id,
-                            name: games[i].name
+                            name: games[i].name,
+                            maxPlayers: games[i].maxPlayers,
+                            curPlayers: games[i].players.length
                         })
                     }
                     socket.emit('gameList', JSON.stringify(clientGames));
                 });
             });
 
+            socket.on('joinGame', function(id) {
+                joinGame(socket, id);
+            });
+
             socket.on('createGame', function(gameName) {
                 var newGame = new Game();
                 newGame.name = gameName;
-                newGame.players.push(socket.decoded.user);
                 newGame.save()
                 .then((result)=>{
-                    var clientGame = {
-                        id: result._id,
-                        name: result.name
-                    }
-                    socket.emit('createdGame', clientGame);
-                    // console.log(result);
+                    joinGame(socket, result._id);
                 })
                 .catch((err)=>{
                     console.log(err);
@@ -125,6 +130,9 @@ module.exports = function(server) {
         //Leave old room
         User.findById(socket.id).exec()
         .then( (user) => {
+            if(!user) {
+                throw new Error('No user found');
+            }
             if(room === user.data.curRoom) {
                 //What to do here?
                 //throw new Error('User is already in that room');
@@ -187,7 +195,78 @@ module.exports = function(server) {
             });
         })
         .catch( (err) => {
-            console.error(err);
+            var errMsg = {
+                title: 'Error joining room',
+                error: err.message
+            };
+            socket.emit('customError', JSON.stringify(errMsg));
+        });
+    }
+
+    function joinGame(socket, gameId) {
+        var foundGame = Game.findById(gameId).exec();
+        var foundUser = User.findById(socket.id).exec();
+
+        Promise.all([foundGame, foundUser])
+        .then( (results) => {
+            var game = results[0];
+            var user = results[1];
+            if(game.players.length === game.maxPlayers) {
+                throw new Error('Game is already full');
+            }
+            if(game.players.indexOf(user._id) !== -1) {
+                throw new Error('You are already in the game');
+            }
+            if(user.data.curGame) {
+                //TODO User is already in another game, confirm modal?
+            }
+
+            user.data.curGame = game._id;
+            game.players.push(user._id);
+
+            var gameSave = game.save();
+            var userSave = user.save();
+            return Promise.all([gameSave, userSave]);
+        })
+        .then( (results) => {
+            socket.emit('joinedGame', results[0]);
+        })
+        .catch((error) => {
+            var errMsg = {
+                title: 'Error joining game',
+                error: error.message
+            };
+            socket.emit('customError', JSON.stringify(errMsg));
+        });
+    }
+
+    function leaveGame(socket, id) {
+        var foundUser = User.findById(socket.id).exec();
+        var foundGame = Game.findById(id).exec();
+
+        Promise.all([foundGame, foundUser])
+        .then( (results) => {
+            var game = results[0];
+            var user = results[1];
+
+            var indexOfUser = game.players.indexOf(user._id);
+            game.players.slice(indexOfUser, 1);
+            user.data.curGame = undefined;
+
+            var gameSave = game.save();
+            var userSave = user.save();
+            return Promise.all([gameSave, userSave]);
+        })
+        .then( (results) => {
+            socket.emit('leftGame', null);
+            console.log(results[1].username + ' left game ' + result[0].name);
+        })
+        .catch((error) => {
+            var errMsg = {
+                title: 'Error leaving game',
+                error: error.message
+            };
+            socket.emit('customError', JSON.stringify(errMsg));
         });
     }
 }
